cloud_functions/summary_generation/requirements.txt
google-cloud-storage
google-cloud-pubsub
google-cloud-bigquery
google-genai
vertexai
requests